pyproject.toml
[project]
name = "ai-request"
version = "0.1.0"
description = ""
authors = [
    {name = "Timothe Duquesne",email = "timothe.duquesne@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
